use template to do registry
